refactor: migrate Accordion components to Vue files and remove TypeScript files
