+++ packages/core/src/accordion/AccordionContent.vue
@@ -1,7 +1,61 @@
+<script lang="ts">
+import type { Ref } from 'vue'
+import type { RadixPrimitiveGetAttrs, RadixPrimitiveReturns } from '../shared/index.ts'
+import {
+  type CollapsibleContentProps,
+  DEFAULT_COLLAPSIBLE_CONTENT_PROPS,
+  useCollapsibleContent,
+  type UseCollapsibleContentProps,
+} from '../collapsible/index.ts'
+import {
+  convertPropsToHookProps,
+  mergePrimitiveAttrs,
+  normalizeAttrs,
+} from '../shared/index.ts'
+import { useAccordionItemContext } from './AccordionItem.vue'
+import { useAccordionContext } from './AccordionRoot.vue'
+
+export type AccordionContentProps = CollapsibleContentProps
+
+export const DEFAULT_ACCORDION_CONTENT_PROPS = DEFAULT_COLLAPSIBLE_CONTENT_PROPS
+
+export interface UseAccordionContentProps extends UseCollapsibleContentProps {}
+
+export function useAccordionContent(props: UseAccordionContentProps): RadixPrimitiveReturns<{
+  isOpen: Ref<boolean>
+  attrs: RadixPrimitiveGetAttrs
+}> {
+  const collapsibleContent = useCollapsibleContent(props)
+
+  const accordionContext = useAccordionContext('AccordionContent')
+  const itemContext = useAccordionItemContext('AccordionContent')
+
+  const style = {
+    '--radix-accordion-content-height': 'var(--radix-collapsible-content-height)',
+    '--radix-accordion-content-width': 'var(--radix-collapsible-content-width)',
+  }
+
+  return {
+    isOpen: collapsibleContent.isOpen,
+    attrs(extraAttrs = []) {
+      const attrs = {
+        'role': 'region',
+        'aria-labelledby': itemContext.triggerId,
+        'data-orientation': accordionContext.orientation,
+        'style': style,
+      }
+
+      mergePrimitiveAttrs(attrs, [collapsibleContent.attrs(), ...extraAttrs])
+
+      return attrs
+    },
+  }
+}
+
+</script>
+
 <script setup lang="ts">
 import { Primitive } from '../primitive/index.ts'
-import { convertPropsToHookProps, normalizeAttrs } from '../shared/index.ts'
-import { type AccordionContentProps, DEFAULT_ACCORDION_CONTENT_PROPS, useAccordionContent } from './AccordionContent.ts'
 
 defineOptions({
   name: 'AccordionContent',
